+++ entity/src/main/java/org/grnet/cat/entities/Setting.java
@@ -1,0 +1,36 @@
+package org.grnet.cat.entities;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.sql.Timestamp;
+
+@Entity
+@Table(name = "t_Setting")
+@Getter
+@Setter
+public class Setting {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "setting_key")
+    private String key;
+
+    @Column(name = "setting_value")
+    private String value;
+
+    @Column(name = "setting_label")
+    private String label;
+
+    @Column(name = "setting_enable")
+    private boolean enabled;
+
+    @Column(name = "updated_on")
+    private Timestamp updatedOn;
+
+    @Column(name = "updated_by")
+    private String updatedBy;
+}